requirements.txt
PyYAML==6.0.2
python-telegram-bot==21.7
requests==2.32.3
bs4==0.0.2
apscheduler==3.11.0
qrcode==8.0
Pillow==11.2.1
alibabacloud_oss_v2==1.1.2
telethon==1.40.0